Rename isExcludedCiper to getFileInfo.
